+++ lib/services/sync-service.ts
@@ -1,0 +1,71 @@
+import { trySyncRepository } from "@/lib/repository";
+
+// Types
+export type SyncResult = { success: boolean; message: string; details?: string };
+
+// Background sync status tracking
+let lastSyncTime: Date | null = null;
+let syncInProgress = false;
+let syncPromise: Promise<SyncResult> | null = null;
+
+// Background sync function
+async function performBackgroundSync(): Promise<SyncResult> {
+  const startTime = Date.now();
+  
+  try {
+    console.log("[SYNC_SERVICE] Starting background repository sync");
+    
+    // Perform the actual sync
+    await trySyncRepository();
+    
+    // Update sync status
+    lastSyncTime = new Date();
+    const duration = Date.now() - startTime;
+    
+    console.log(`[SYNC_SERVICE] Background sync completed successfully in ${duration}ms`);
+    
+    return {
+      success: true,
+      message: "Repository synchronized successfully",
+      details: `Sync completed in ${duration}ms`,
+    };
+  } catch (error) {
+    const duration = Date.now() - startTime;
+    const errorMessage = error instanceof Error ? error.message : "Unknown sync error";
+    
+    console.error(`[SYNC_SERVICE] Background sync failed after ${duration}ms:`, error);
+    
+    return {
+      success: false,
+      message: "Repository synchronization failed",
+      details: `Sync failed after ${duration}ms: ${errorMessage}`,
+    };
+  }
+}
+
+// Helper function to start automatic background sync
+export async function startBackgroundSync(): Promise<SyncResult | null> {
+  if (syncInProgress) {
+    console.log("[SYNC_SERVICE] Background sync already in progress");
+    return null;
+  }
+
+  console.log("[SYNC_SERVICE] Starting automatic background sync");
+  
+  syncInProgress = true;
+  syncPromise = performBackgroundSync().finally(() => {
+    syncInProgress = false;
+    syncPromise = null;
+  });
+
+  return await syncPromise;
+}
+
+// Export sync status for other modules
+export function getSyncStatus() {
+  return {
+    syncInProgress,
+    lastSyncTime,
+    timeSinceLastSync: lastSyncTime ? Date.now() - lastSyncTime.getTime() : null,
+  };
+}
